The aweek class

The method c.aweek() will always return an aweek object with the same week_start attribute as the first object in the list. If the first object is also a factor, then the output will be re-leveled. If week-like characters are presented (e.g. "2019-W10-1"), then these are assumed to have the same week_start as the first object.

# S3 method for aweek
print(x, ...)

# S3 method for aweek
[(x, i)

# S3 method for aweek
c(..., recursive = FALSE, use.names = TRUE)

Arguments

x
an object of class aweek

...
a series of aweek objects (unused in print.aweek())

i
index for subsetting an aweek object.

recursive, use.names
parameters passed on to unlist()

Value

an object of class aweek

Details

The aweek class is a character or factor in the format YYYY-Www(-d) with a "week_start" attribute containing an integer specifying which day of the ISO 8601 week each week should begin. This documentation shows how to print or subset the object.

See also

Examples

d <- as.Date("2018-12-20") + 1:40
w <- date2week(d, week_start = "Sunday")
print(w)

#> <aweek start: Sunday>
#>  [1] "2018-W51-6" "2018-W51-7" "2018-W52-1" "2018-W52-2" "2018-W52-3"
#>  [6] "2018-W52-4" "2018-W52-5" "2018-W52-6" "2018-W52-7" "2019-W01-1"
#> [11] "2019-W01-2" "2019-W01-3" "2019-W01-4" "2019-W01-5" "2019-W01-6"
#> [16] "2019-W01-7" "2019-W02-1" "2019-W02-2" "2019-W02-3" "2019-W02-4"
#> [21] "2019-W02-5" "2019-W02-6" "2019-W02-7" "2019-W03-1" "2019-W03-2"
#> [26] "2019-W03-3" "2019-W03-4" "2019-W03-5" "2019-W03-6" "2019-W03-7"
#> [31] "2019-W04-1" "2019-W04-2" "2019-W04-3" "2019-W04-4" "2019-W04-5"
#> [36] "2019-W04-6" "2019-W04-7" "2019-W05-1" "2019-W05-2" "2019-W05-3"

# subsetting acts as normal
w[1:10]

#> <aweek start: Sunday>
#>  [1] "2018-W51-6" "2018-W51-7" "2018-W52-1" "2018-W52-2" "2018-W52-3"
#>  [6] "2018-W52-4" "2018-W52-5" "2018-W52-6" "2018-W52-7" "2019-W01-1"

# Combining multiple aweek objects
c(w[1], w[3], w[5])

#> <aweek start: Sunday>
#> [1] "2018-W51-6" "2018-W52-1" "2018-W52-3"

# differing week_start days will default to the attribute of the first 
# aweek object
mon <- date2week(Sys.Date(), week_start = "Monday")
mon

#> <aweek start: Monday>
#> [1] "2019-W12-1"

c(w, mon)

#> <aweek start: Sunday>
#>  [1] "2018-W51-6" "2018-W51-7" "2018-W52-1" "2018-W52-2" "2018-W52-3"
#>  [6] "2018-W52-4" "2018-W52-5" "2018-W52-6" "2018-W52-7" "2019-W01-1"
#> [11] "2019-W01-2" "2019-W01-3" "2019-W01-4" "2019-W01-5" "2019-W01-6"
#> [16] "2019-W01-7" "2019-W02-1" "2019-W02-2" "2019-W02-3" "2019-W02-4"
#> [21] "2019-W02-5" "2019-W02-6" "2019-W02-7" "2019-W03-1" "2019-W03-2"
#> [26] "2019-W03-3" "2019-W03-4" "2019-W03-5" "2019-W03-6" "2019-W03-7"
#> [31] "2019-W04-1" "2019-W04-2" "2019-W04-3" "2019-W04-4" "2019-W04-5"
#> [36] "2019-W04-6" "2019-W04-7" "2019-W05-1" "2019-W05-2" "2019-W05-3"
#> [41] "2019-W12-2"

c(mon, w)

#> <aweek start: Monday>
#>  [1] "2019-W12-1" "2018-W51-5" "2018-W51-6" "2018-W51-7" "2018-W52-1"
#>  [6] "2018-W52-2" "2018-W52-3" "2018-W52-4" "2018-W52-5" "2018-W52-6"
#> [11] "2018-W52-7" "2019-W01-1" "2019-W01-2" "2019-W01-3" "2019-W01-4"
#> [16] "2019-W01-5" "2019-W01-6" "2019-W01-7" "2019-W02-1" "2019-W02-2"
#> [21] "2019-W02-3" "2019-W02-4" "2019-W02-5" "2019-W02-6" "2019-W02-7"
#> [26] "2019-W03-1" "2019-W03-2" "2019-W03-3" "2019-W03-4" "2019-W03-5"
#> [31] "2019-W03-6" "2019-W03-7" "2019-W04-1" "2019-W04-2" "2019-W04-3"
#> [36] "2019-W04-4" "2019-W04-5" "2019-W04-6" "2019-W04-7" "2019-W05-1"
#> [41] "2019-W05-2"

# combining Dates will be coerced to aweek objects under the same rules
c(w, Sys.Date())

#> <aweek start: Sunday>
#>  [1] "2018-W51-6" "2018-W51-7" "2018-W52-1" "2018-W52-2" "2018-W52-3"
#>  [6] "2018-W52-4" "2018-W52-5" "2018-W52-6" "2018-W52-7" "2019-W01-1"
#> [11] "2019-W01-2" "2019-W01-3" "2019-W01-4" "2019-W01-5" "2019-W01-6"
#> [16] "2019-W01-7" "2019-W02-1" "2019-W02-2" "2019-W02-3" "2019-W02-4"
#> [21] "2019-W02-5" "2019-W02-6" "2019-W02-7" "2019-W03-1" "2019-W03-2"
#> [26] "2019-W03-3" "2019-W03-4" "2019-W03-5" "2019-W03-6" "2019-W03-7"
#> [31] "2019-W04-1" "2019-W04-2" "2019-W04-3" "2019-W04-4" "2019-W04-5"
#> [36] "2019-W04-6" "2019-W04-7" "2019-W05-1" "2019-W05-2" "2019-W05-3"
#> [41] "2019-W12-2"

# truncated aweek objects will be un-truncated
w2 <- date2week(d[1:5], week_start = "Monday", floor_day = TRUE)
w2

#> <aweek start: Monday>
#> [1] "2018-W51" "2018-W51" "2018-W51" "2018-W52" "2018-W52"

c(w[1:5], w2)

#> <aweek start: Sunday>
#>  [1] "2018-W51-6" "2018-W51-7" "2018-W52-1" "2018-W52-2" "2018-W52-3"
#>  [6] "2018-W51-2" "2018-W51-2" "2018-W51-2" "2018-W52-2" "2018-W52-2"