Convert date to a an arbitrary week definition

Convert date to a an arbitrary week definition

date2week(x, week_start = 1, floor_day = FALSE, numeric = FALSE,
  factor = FALSE, ...)

week2date(x, week_start = 1, floor_day = FALSE)

Arguments

x
a Date, POSIXt, character, or any data that can be easily converted to a date with as.POSIXlt().

week_start
a number indicating the start of the week based on the ISO 8601 standard from 1 to 7 where 1 = Monday OR an abbreviation of the weekdate in an English or current locale. Note: using a non-English locale may render your code non-portable.

floor_day
when TRUE, the days will be set to the start of the week.

numeric
if TRUE, only the numeric week be returned. If FALSE (default), the date in the format "YYYY-Www-d" will be returned.

factor
if TRUE, a factor will be returned with levels spanning the range of dates. If floor_date = FALSE, then this will use the sequence of days between the first and last date, but if floor_date = TRUE, then the sequence of weeks between the first and last date will be used. Take caution when using this with a large date range as the resulting factor can contain all days between dates.

...
arguments passed to as.POSIXlt(), unused in all other cases.

Details

Weeks differ in their start dates depending on context. The ISO 8601 standard specifies that Monday starts the week (https://en.wikipedia.org/wiki/ISO_week_date) while the US CDC uses Sunday as the start of the week (https://wwwn.cdc.gov/nndss/document/MMWR_Week_overview.pdf). For example, MSF has varying start dates depending on country in order to better coordinate response.

While there are packages that provide conversion for ISOweeks and epiweeks, these do not provide seamless conversion from dates to epiweeks with non-standard start dates. This package provides a lightweight utility to be able to convert each day.

Note

date2week() will initially convert the input with as.POSIXlt() and use that to calculate the week. If the user supplies character input, it is expected that the input will be of the format yyyy-mm-dd unless the user explicitly passes the "format" parameter to as.POSIXlt(). If the input is not in yyyy-mm-dd and the format parameter is not passed, date2week() will result in an error.

See also

Examples

# The same set of days will occur in different weeks depending on the start
# date. Here we can define a week before and after today

print(dat <- Sys.Date() + -6:7)

#>  [1] "2019-04-18" "2019-04-19" "2019-04-20" "2019-04-21" "2019-04-22"
#>  [6] "2019-04-23" "2019-04-24" "2019-04-25" "2019-04-26" "2019-04-27"
#> [11] "2019-04-28" "2019-04-29" "2019-04-30" "2019-05-01"

# By default, the weeks are defined as ISO weeks, which start on Monday
print(iso_dat <- date2week(dat))

#> <aweek start: Monday>
#>  [1] "2019-W16-4" "2019-W16-5" "2019-W16-6" "2019-W16-7" "2019-W17-1"
#>  [6] "2019-W17-2" "2019-W17-3" "2019-W17-4" "2019-W17-5" "2019-W17-6"
#> [11] "2019-W17-7" "2019-W18-1" "2019-W18-2" "2019-W18-3"

# If you want lubridate-style numeric-only weeks, you need look no further
# than the "numeric" argument
date2week(dat, 1, numeric = TRUE)

#>  [1] 16 16 16 16 17 17 17 17 17 17 17 18 18 18

# You can also convert to factor and include all of the missing dates, but
# beware that this may result in a very large factor due to the number of
# levels present
date2week(Sys.Date() + c(0, 10), factor = TRUE)

#> <aweek start: Monday>
#> [1] 2019-W17-3 2019-W18-6
#> 11 Levels: 2019-W17-3 2019-W17-4 2019-W17-5 2019-W17-6 ... 2019-W18-6

# The aweek class can be converted back to a date with `as.Date()`
as.Date(iso_dat)

#>  [1] "2019-04-18" "2019-04-19" "2019-04-20" "2019-04-21" "2019-04-22"
#>  [6] "2019-04-23" "2019-04-24" "2019-04-25" "2019-04-26" "2019-04-27"
#> [11] "2019-04-28" "2019-04-29" "2019-04-30" "2019-05-01"

# If you want to show only the first day of the week, you can use the 
# `floor_day` argument
as.Date(iso_dat, floor_day = TRUE)

#>  [1] "2019-04-15" "2019-04-15" "2019-04-15" "2019-04-15" "2019-04-22"
#>  [6] "2019-04-22" "2019-04-22" "2019-04-22" "2019-04-22" "2019-04-22"
#> [11] "2019-04-22" "2019-04-29" "2019-04-29" "2019-04-29"

# This also works with `factor`:
as.Date(iso_dat, floor_day = TRUE, factor = TRUE)

#>  [1] "2019-04-15" "2019-04-15" "2019-04-15" "2019-04-15" "2019-04-22"
#>  [6] "2019-04-22" "2019-04-22" "2019-04-22" "2019-04-22" "2019-04-22"
#> [11] "2019-04-22" "2019-04-29" "2019-04-29" "2019-04-29"

# ISO week definition: Monday -- 1
date2week(dat, 1)

#> <aweek start: Monday>
#>  [1] "2019-W16-4" "2019-W16-5" "2019-W16-6" "2019-W16-7" "2019-W17-1"
#>  [6] "2019-W17-2" "2019-W17-3" "2019-W17-4" "2019-W17-5" "2019-W17-6"
#> [11] "2019-W17-7" "2019-W18-1" "2019-W18-2" "2019-W18-3"

date2week(dat, "Monday")

#> <aweek start: Monday>
#>  [1] "2019-W16-4" "2019-W16-5" "2019-W16-6" "2019-W16-7" "2019-W17-1"
#>  [6] "2019-W17-2" "2019-W17-3" "2019-W17-4" "2019-W17-5" "2019-W17-6"
#> [11] "2019-W17-7" "2019-W18-1" "2019-W18-2" "2019-W18-3"

# Tuesday -- 2
date2week(dat, 2)

#> <aweek start: Tuesday>
#>  [1] "2019-W16-3" "2019-W16-4" "2019-W16-5" "2019-W16-6" "2019-W16-7"
#>  [6] "2019-W17-1" "2019-W17-2" "2019-W17-3" "2019-W17-4" "2019-W17-5"
#> [11] "2019-W17-6" "2019-W17-7" "2019-W18-1" "2019-W18-2"

date2week(dat, "Tuesday")

#> <aweek start: Tuesday>
#>  [1] "2019-W16-3" "2019-W16-4" "2019-W16-5" "2019-W16-6" "2019-W16-7"
#>  [6] "2019-W17-1" "2019-W17-2" "2019-W17-3" "2019-W17-4" "2019-W17-5"
#> [11] "2019-W17-6" "2019-W17-7" "2019-W18-1" "2019-W18-2"

# Wednesday -- 3
date2week(dat, 3)

#> <aweek start: Wednesday>
#>  [1] "2019-W16-2" "2019-W16-3" "2019-W16-4" "2019-W16-5" "2019-W16-6"
#>  [6] "2019-W16-7" "2019-W17-1" "2019-W17-2" "2019-W17-3" "2019-W17-4"
#> [11] "2019-W17-5" "2019-W17-6" "2019-W17-7" "2019-W18-1"

date2week(dat, "W") # you can use valid abbreviations

#> <aweek start: Wednesday>
#>  [1] "2019-W16-2" "2019-W16-3" "2019-W16-4" "2019-W16-5" "2019-W16-6"
#>  [6] "2019-W16-7" "2019-W17-1" "2019-W17-2" "2019-W17-3" "2019-W17-4"
#> [11] "2019-W17-5" "2019-W17-6" "2019-W17-7" "2019-W18-1"

# Thursday -- 4
date2week(dat, 4)

#> <aweek start: Thursday>
#>  [1] "2019-W16-1" "2019-W16-2" "2019-W16-3" "2019-W16-4" "2019-W16-5"
#>  [6] "2019-W16-6" "2019-W16-7" "2019-W17-1" "2019-W17-2" "2019-W17-3"
#> [11] "2019-W17-4" "2019-W17-5" "2019-W17-6" "2019-W17-7"

date2week(dat, "Thursday")

#> <aweek start: Thursday>
#>  [1] "2019-W16-1" "2019-W16-2" "2019-W16-3" "2019-W16-4" "2019-W16-5"
#>  [6] "2019-W16-6" "2019-W16-7" "2019-W17-1" "2019-W17-2" "2019-W17-3"
#> [11] "2019-W17-4" "2019-W17-5" "2019-W17-6" "2019-W17-7"

# Friday -- 5
date2week(dat, 5)

#> <aweek start: Friday>
#>  [1] "2019-W15-7" "2019-W16-1" "2019-W16-2" "2019-W16-3" "2019-W16-4"
#>  [6] "2019-W16-5" "2019-W16-6" "2019-W16-7" "2019-W17-1" "2019-W17-2"
#> [11] "2019-W17-3" "2019-W17-4" "2019-W17-5" "2019-W17-6"

date2week(dat, "Friday")

#> <aweek start: Friday>
#>  [1] "2019-W15-7" "2019-W16-1" "2019-W16-2" "2019-W16-3" "2019-W16-4"
#>  [6] "2019-W16-5" "2019-W16-6" "2019-W16-7" "2019-W17-1" "2019-W17-2"
#> [11] "2019-W17-3" "2019-W17-4" "2019-W17-5" "2019-W17-6"

# Saturday -- 6
date2week(dat, 6)

#> <aweek start: Saturday>
#>  [1] "2019-W16-6" "2019-W16-7" "2019-W17-1" "2019-W17-2" "2019-W17-3"
#>  [6] "2019-W17-4" "2019-W17-5" "2019-W17-6" "2019-W17-7" "2019-W18-1"
#> [11] "2019-W18-2" "2019-W18-3" "2019-W18-4" "2019-W18-5"

date2week(dat, "Saturday")

#> <aweek start: Saturday>
#>  [1] "2019-W16-6" "2019-W16-7" "2019-W17-1" "2019-W17-2" "2019-W17-3"
#>  [6] "2019-W17-4" "2019-W17-5" "2019-W17-6" "2019-W17-7" "2019-W18-1"
#> [11] "2019-W18-2" "2019-W18-3" "2019-W18-4" "2019-W18-5"

# Epiweek definition: Sunday -- 7 
date2week(dat, 7)

#> <aweek start: Sunday>
#>  [1] "2019-W16-5" "2019-W16-6" "2019-W16-7" "2019-W17-1" "2019-W17-2"
#>  [6] "2019-W17-3" "2019-W17-4" "2019-W17-5" "2019-W17-6" "2019-W17-7"
#> [11] "2019-W18-1" "2019-W18-2" "2019-W18-3" "2019-W18-4"

date2week(dat, "Sunday")

#> <aweek start: Sunday>
#>  [1] "2019-W16-5" "2019-W16-6" "2019-W16-7" "2019-W17-1" "2019-W17-2"
#>  [6] "2019-W17-3" "2019-W17-4" "2019-W17-5" "2019-W17-6" "2019-W17-7"
#> [11] "2019-W18-1" "2019-W18-2" "2019-W18-3" "2019-W18-4"